refactor(archive): 将7z和tar归档功能从base模块分离到独立子模块
- 创建archive_7z子模块，包含7zArchive.cpp实现文件
- 创建archive_tar子模块，包含TARArchive.cpp和tar.h头文件
- 从base模块移除7zArchive.cpp和TARArchive.cpp源文件
- 添加ArchiveFilename.cpp/h用于归档文件名处理
- 更新CMakeLists.txt配置归档模块依赖关系
- 修改文档CORE_MODULE_DECOUPLING.md更新模块拆分状态
