+++ cpp/core/archive/CMakeLists.txt
@@ -3,3 +3,5 @@ project(core_archive LANGUAGES CXX)
 
 add_subdirectory(xp3)
 add_subdirectory(zip)
+add_subdirectory(7z)
+add_subdirectory(tar)
