+++ cpp/core/base/CMakeLists.txt
@@ -3,14 +3,12 @@ project(core_base_module LANGUAGES CXX)
 
 set(BASE_PATH ${CMAKE_CURRENT_LIST_DIR})
 set(BASE_SOURCE_FILES
-    ${BASE_PATH}/7zArchive.cpp
     ${BASE_PATH}/BinaryStream.cpp
     ${BASE_PATH}/CharacterSet.cpp
     ${BASE_PATH}/EventIntf.cpp
     ${BASE_PATH}/ScriptMgnIntf.cpp
     ${BASE_PATH}/SysInitIntf.cpp
     ${BASE_PATH}/SystemIntf.cpp
-    ${BASE_PATH}/TARArchive.cpp
     ${BASE_PATH}/TextStream.cpp
     ${BASE_PATH}/KAGParser.cpp
     ${BASE_PATH}/MsgIntf.cpp
@@ -35,7 +33,7 @@ set(BASE_HEADERS_DIR
 
 add_library(${PROJECT_NAME} STATIC ${BASE_SOURCE_FILES})
 target_include_directories(${PROJECT_NAME} PUBLIC ${BASE_HEADERS_DIR})
-target_link_libraries(${PROJECT_NAME} PUBLIC tjs2 archive_xp3 archive_zip core_io core_storage PRIVATE core_visual_module core_plugin_module core_environ_module core_extension_module core_sound_module core_utils_module)
+target_link_libraries(${PROJECT_NAME} PUBLIC tjs2 archive_xp3 archive_zip archive_7z archive_tar core_io core_storage PRIVATE core_visual_module core_plugin_module core_environ_module core_extension_module core_sound_module core_utils_module)
 
 find_package(uchardet CONFIG REQUIRED)
 find_package(unrar CONFIG REQUIRED)
