+++ cpp/core/storage/impl/ArchiveFilename.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "StorageIntf.h"
+
+void storeFilename(ttstr &name, const char *narrowName, const ttstr &filename);
